# adapters/src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Price product API
  version: 1.0.0
  description: API to query product price rates based on date, product ID, and brand ID.

servers:
  - url: http://localhost:8080
    description: Production server

paths:
  /products:
    get:
      summary: Get applicable price rate
      operationId: getApplicableRate
      parameters:
        - name: applicationDate
          in: query
          description: Date on which the rate is to be applied (YYYY-MM-DD format).
          required: true
          schema:
            type: string
            format: date
            example: "2025-06-10"
        - name: productId
          in: query
          description: Unique identifier of the product.
          required: true
          schema:
            type: integer
            format: int64
            example: 35455
        - name: brandId
          in: query
          description: Unique identifier of the commercial brand.
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: Price rate successfully retrieved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppliedRate"
        "400":
          description: Invalid input parameters.
        "404":
          description: Rate not found for the provided parameters.
        "500":
          description: Internal server error.

components:
  schemas:
    AppliedRate:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          description: Unique identifier of the product.
          example: 35455
        brandId:
          type: integer
          format: int64
          description: Unique identifier of the commercial brand.
          example: 1
        rateToApply:
          type: string
          description: Code or identifier of the rate to be applied.
          example: "GENERAL_RATE_1"
        applicationStartDate:
          type: string
          format: date-time
          description: Start date and time of the rate's validity.
          example: "2025-06-10T00:00:00Z"
        applicationEndDate:
          type: string
          format: date-time
          description: End date and time of the rate's validity.
          example: "2025-12-31T23:59:59Z"
        finalPriceToApply:
          type: number
          format: float
          description: Final price to be applied to the product.
          example: 35.50
      required:
        - productId
        - brandId
        - rateToApply
        - applicationStartDate
        - applicationEndDate
        - finalPriceToApply